requirements.txt
django
mysqlclient
djangorestframework
segno
